requirements.txt
openai==0.27.2
streamlit==1.25.0
snowflake-connector-python==3.1.0
plotly.express==0.4.1
altair==4.2.2 
streamlit_option_menu==0.3.6 
pandas==1.4.3 
snowflake-snowpark-python==1.6.1
pyarrow==10.0.1
pyperclip==1.8.2
